BufferedDataIO::Flush(), use ssize_t rather than size_t
Write() was returing an ssize_t and Flush() was expecting a ssize_t,
but a size_t was instead used. CID 11052.

+++ src/kits/support/BufferedDataIO.cpp
@@ -86,8 +86,8 @@ BBufferedDataIO::Flush()
 	if (!fDirty)
 		return B_OK;
 
-	size_t bytesWritten = fStream.Write(fBuffer + fPosition, fSize);
-	if (bytesWritten == fSize) {
+	ssize_t bytesWritten = fStream.Write(fBuffer + fPosition, fSize);
+	if ((size_t)bytesWritten == fSize) {
 		fDirty = false;
 		fPosition = 0;
 		fSize = 0;
